+++ src/viewer/icons/index.ts
@@ -1,5 +1,7 @@
 import CollapseIcon from "./collapse.svg";
 import ExpandIcon from "./expand.svg";
+import PinIcon from "./pin.svg";
+import UnpinIcon from "./unpin.svg";
 import CopyDeckIcon from "./copy_deck.svg";
 
-export { CollapseIcon, ExpandIcon, CopyDeckIcon };
+export { CollapseIcon, ExpandIcon, PinIcon, UnpinIcon, CopyDeckIcon };